feat: parse attributes (#1)
* feat: added parsing of attributes (inputs) for module calls

This adds parsing of attributes (inputs) to module calls. Attributes are exposed as a new 'Attributes' field on the 'ModuleCall' struct.

* feat: traversal attributes

Attributes that are traversal types (such as variables and data sources) are now included in parsed module calls.

* chore: refactor module

This refactors the structure of the module so the root package contains all the importable logic. It also renames everything to the new module name.

* fix: interpolated strings

* ci: various improvements for CI

* chore: README updates

* feat: attributes for resource blocks

* ci: clean up ci jobs

* chore: example for load module

* chore: README update

* chore: docs

* ci: various fixes